fix(checkout): transactions after switching tabs was not consistent
- Introduce a new key in Transaction via bit packing to represent sale id and product id together
- Used the bit packed key as the lookup in TransactionStore

+++ Mhyrenz_Interface/Store/ITransactionStore.cs
@@ -6,7 +6,7 @@ using Mhyrenz_Interface.Features.Checkout.ViewModels;
 
 namespace Mhyrenz_Interface.Store
 {
-    public interface ITransactionStore : IViewModelStore<int, TransactionDataViewModel>
+    public interface ITransactionStore : IViewModelStore<long, TransactionDataViewModel>
     {
         event EventHandler<Sale> SaleChange;
